refactor(types): rename sidebar-item file and update its import
