+++ src/types/SidebarItem.ts
