+++ src/types/index.ts
@@ -2,3 +2,4 @@ export * from './AuthFormField';
 export * from './AlertLevel';
 export * from './Tool';
 export * from './SortOption';
+export * from './SidebarItem';
